requirements.txt
discord.py
aiosqlite
PyNaCl
python-dotenv
